refactor: introduced Go package structure
Refactored code from main.go to internal folders to follow Go's guidelines on
style and structure.

The following GitHub contains more detail: github.com/golang-standards/project-layout
